firmware: add crunchdefender (STM32F405) target and fix STM32 builds
Brings the crunchdefender dual-USB mount limit guard into the build and
repairs the STM32 firmware builds, which were red in CI.

CI was failing because the pinned ST HAL drivers are newer than the
hand-written config headers, plus the crunchdefender target referenced a
vendor tree that was never registered:

- Register STM32CubeF4 as a submodule. common/CMakeLists.txt already builds
  the F4 HAL gated on this tree existing, and the README/CI assumed it was
  present, but it was missing from .gitmodules so `submodules: recursive`
  never fetched it.

- Build only the selected MCU family's HAL. The common library gated each
  HAL on the vendor tree merely existing, so once STM32CubeF4 was present a
  G0 build also tried to compile the F4 HAL (and vice versa) without the
  per-project hal_conf.h on the include path. Gate on MCU_FAMILY as well.

- Define EXTERNAL_I2S1/2_CLOCK_VALUE (G0) and EXTERNAL_CLOCK_VALUE (F4) in
  the project hal_conf.h files. The pinned RCC HAL references these I2S
  external-clock macros unconditionally; values match ST's hal_conf templates.

- Declare AHBPrescTable/APBPrescTable as uint32_t in the G0 system file to
  match the CMSIS G0 device header (the F4 CMSIS headers use uint8_t).

- Emit <project>.elf via the executable SUFFIX so CI's build-info and
  artifact-upload steps (which reference .elf) succeed now that the build
  gets that far.

Both pulsardew (G0B1) and crunchdefender (F405) now configure, build, and
link cleanly, and the CI build-info step passes.

+++ firmware/pulsardew/src/system_stm32g0xx.c
@@ -1,8 +1,10 @@
 #include "stm32g0xx.h"
 
 uint32_t SystemCoreClock = 64000000UL;
-const uint8_t AHBPrescTable[16] = {0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 6, 7, 8, 9};
-const uint8_t APBPrescTable[8] = {0, 0, 0, 0, 1, 2, 3, 4};
+/* uint32_t to match the extern declarations in the CMSIS G0 device header
+ * (system_stm32g0xx.h); the F4 CMSIS headers declare these as uint8_t. */
+const uint32_t AHBPrescTable[16] = {0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 6, 7, 8, 9};
+const uint32_t APBPrescTable[8] = {0, 0, 0, 0, 1, 2, 3, 4};
 
 void SystemInit(void)
 {
